Fixed float error

+++ calculator.js
@@ -10,32 +10,32 @@ $(function() {
 	
 	$("#add").on("click",function(){
 		//input 1 + input 2
-		var num1 = parseInt($("#inputNum1").val());
-		var num2 = parseInt($("#inputNum2").val());
+		var num1 = parseFloat($("#inputNum1").val());
+		var num2 = parseFloat($("#inputNum2").val());
 		var sum = num1 + num2;
 		console.log(sum);
 	});	
 
 	$("#subtract").on("click",function(){
 		//input 1 + input 2
-		var num1 = parseInt($("#inputNum1").val());
-		var num2 = parseInt($("#inputNum2").val());
+		var num1 = parseFloat($("#inputNum1").val());
+		var num2 = parseFloat($("#inputNum2").val());
 		var subtraction = num1 - num2;
 		console.log(subtraction);
 	})
 
 	$("#multiply").on("click",function(){
 		//input 1 + input 2
-		var num1 = parseInt($("#inputNum1").val());
-		var num2 = parseInt($("#inputNum2").val());
+		var num1 = parseFloat($("#inputNum1").val());
+		var num2 = parseFloat($("#inputNum2").val());
 		var multiply = num1 * num2;
 		console.log(multiply);
 	})
 
 	$("#divide").on("click",function(){
 		//input 1 + input 2
-		var num1 = parseInt($("#inputNum1").val());
-		var num2 = parseInt($("#inputNum2").val());
+		var num1 = parseFloat($("#inputNum1").val());
+		var num2 = parseFloat($("#inputNum2").val());
 		var division = num1/num2;
 		console.log(division);
 	})
